+++ apps/decodex/src/orchestrator/execution.rs
@@ -1,7 +1,4 @@
-use agent::{
-	AppServerThreadArchiveOutcome, AppServerThreadArchiveRequest, CodexAccountAuthFailure,
-	CodexAccountPool, CodexAccountProvider,
-};
+use agent::{CodexAccountAuthFailure, CodexAccountPool, CodexAccountProvider};
 use git_credentials::GitSigningConfig;
 
 struct AgentGitCredentialEnvironment {
@@ -45,25 +42,6 @@ where
 	decodex_tool_bridge: &'a DecodexToolBridge<'a>,
 	phase_goal_controller: &'a dyn PhaseGoalController,
 	codex_account_provider: Option<&'a dyn CodexAccountProvider>,
-}
-
-#[derive(Clone)]
-struct ThreadArchiveCandidate {
-	issue_id: String,
-	issue_identifier: String,
-	run_id: String,
-	attempt_number: i64,
-	thread_id: String,
-	sequence_number: i64,
-}
-
-struct ThreadArchiveCandidateSource<'a> {
-	run_id: &'a str,
-	issue_id: &'a str,
-	issue_identifier: &'a str,
-	attempt_number: i64,
-	thread_id: &'a str,
-	sequence_number: Option<i64>,
 }
 
 enum IssueAppServerRunOutcome {
@@ -527,316 +505,6 @@ where
 	Ok(run_summary_from_issue_run(run.project.service_id(), run.issue_run))
 }
 
-fn archive_completed_issue_threads_best_effort(
-	project: &ServiceConfig,
-	state_store: &StateStore,
-	issue_run: &IssueRunPlan,
-	process_env: &AppServerProcessEnv,
-	transport: &str,
-	run_result: &AppServerRunResult,
-) {
-	let current = ThreadArchiveCandidate {
-		issue_id: issue_run.issue.id.clone(),
-		issue_identifier: issue_run.issue.identifier.clone(),
-		run_id: issue_run.run_id.clone(),
-		attempt_number: issue_run.attempt_number,
-		thread_id: run_result.thread_id.clone(),
-		sequence_number: run_result.event_count.saturating_add(1),
-	};
-
-	archive_issue_threads_best_effort(
-		project,
-		state_store,
-		&issue_run.issue.id,
-		&issue_run.issue.identifier,
-		process_env,
-		transport,
-		Some(current),
-	);
-}
-
-fn reconcile_terminal_thread_archive_backlog_best_effort(
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-) {
-	let process_env = AppServerProcessEnv::default();
-	let transport = workflow.frontmatter().agent().transport();
-	let candidates =
-		match terminal_thread_archive_backlog_candidates(state_store, project.service_id()) {
-			Ok(candidates) => candidates,
-			Err(error) => {
-				tracing::warn!(
-					?error,
-					project_id = project.service_id(),
-					"Failed to list terminal thread archive backlog; skipping this archive reconciliation pass."
-				);
-
-				return;
-			},
-		};
-
-	for candidate in candidates {
-		archive_completed_issue_thread_best_effort(
-			project,
-			state_store,
-			&process_env,
-			transport,
-			&candidate,
-		);
-	}
-}
-
-fn archive_issue_threads_best_effort(
-	project: &ServiceConfig,
-	state_store: &StateStore,
-	issue_id: &str,
-	issue_identifier: &str,
-	process_env: &AppServerProcessEnv,
-	transport: &str,
-	current: Option<ThreadArchiveCandidate>,
-) {
-	let fallback_candidate = current.clone();
-	let candidates =
-		match issue_thread_archive_candidates(state_store, issue_id, issue_identifier, current) {
-			Ok(candidates) => candidates,
-			Err(error) => {
-				tracing::warn!(
-					?error,
-					project_id = project.service_id(),
-					issue_id,
-					issue = issue_identifier,
-					"Failed to list completed issue threads for archive; archiving current thread only."
-				);
-
-				fallback_candidate.into_iter().collect()
-			},
-		};
-
-	for candidate in candidates {
-		archive_completed_issue_thread_best_effort(
-			project,
-			state_store,
-			process_env,
-			transport,
-			&candidate,
-		);
-	}
-}
-
-#[cfg(test)]
-fn completed_issue_thread_archive_candidates(
-	state_store: &StateStore,
-	issue_run: &IssueRunPlan,
-	run_result: &AppServerRunResult,
-) -> Result<Vec<ThreadArchiveCandidate>> {
-	issue_thread_archive_candidates(
-		state_store,
-		&issue_run.issue.id,
-		&issue_run.issue.identifier,
-		Some(ThreadArchiveCandidate {
-			issue_id: issue_run.issue.id.clone(),
-			issue_identifier: issue_run.issue.identifier.clone(),
-			run_id: issue_run.run_id.clone(),
-			attempt_number: issue_run.attempt_number,
-			thread_id: run_result.thread_id.clone(),
-			sequence_number: run_result.event_count.saturating_add(1),
-		}),
-	)
-}
-
-fn issue_thread_archive_candidates(
-	state_store: &StateStore,
-	issue_id: &str,
-	issue_identifier: &str,
-	current: Option<ThreadArchiveCandidate>,
-) -> Result<Vec<ThreadArchiveCandidate>> {
-	let mut seen_thread_ids = HashSet::new();
-	let mut candidates = Vec::new();
-
-	if let Some(current) = current {
-		push_thread_archive_candidate(
-			state_store,
-			&mut candidates,
-			&mut seen_thread_ids,
-			ThreadArchiveCandidateSource {
-				run_id: &current.run_id,
-				issue_id: &current.issue_id,
-				issue_identifier: &current.issue_identifier,
-				attempt_number: current.attempt_number,
-				thread_id: &current.thread_id,
-				sequence_number: Some(current.sequence_number),
-			},
-		)?;
-	}
-
-	for attempt in state_store.list_run_attempts_for_issue(issue_id)? {
-		if !completed_issue_archive_attempt_status(attempt.status()) {
-			continue;
-		}
-
-		if let Some(thread_id) = attempt.thread_id() {
-			push_thread_archive_candidate(
-				state_store,
-				&mut candidates,
-				&mut seen_thread_ids,
-				ThreadArchiveCandidateSource {
-					run_id: attempt.run_id(),
-					issue_id: attempt.issue_id(),
-					issue_identifier,
-					attempt_number: attempt.attempt_number(),
-					thread_id,
-					sequence_number: None,
-				},
-			)?;
-		}
-	}
-
-	Ok(candidates)
-}
-
-fn terminal_thread_archive_backlog_candidates(
-	state_store: &StateStore,
-	project_id: &str,
-) -> Result<Vec<ThreadArchiveCandidate>> {
-	let mut seen_thread_ids = HashSet::new();
-	let mut candidates = Vec::new();
-
-	for attempt in state_store.list_run_attempts_for_project(project_id)? {
-		if !completed_issue_archive_attempt_status(attempt.status()) {
-			continue;
-		}
-
-		if let Some(thread_id) = attempt.thread_id() {
-			push_thread_archive_candidate(
-				state_store,
-				&mut candidates,
-				&mut seen_thread_ids,
-				ThreadArchiveCandidateSource {
-					run_id: attempt.run_id(),
-					issue_id: attempt.issue_id(),
-					issue_identifier: attempt.issue_id(),
-					attempt_number: attempt.attempt_number(),
-					thread_id,
-					sequence_number: None,
-				},
-			)?;
-		}
-	}
-
-	Ok(candidates)
-}
-
-fn completed_issue_archive_attempt_status(status: &str) -> bool {
-	matches!(
-		status,
-		"succeeded" | "failed" | "interrupted" | "terminated" | TERMINAL_GUARDED_RUN_STATUS
-	)
-}
-
-fn push_thread_archive_candidate(
-	state_store: &StateStore,
-	candidates: &mut Vec<ThreadArchiveCandidate>,
-	seen_thread_ids: &mut HashSet<String>,
-	source: ThreadArchiveCandidateSource<'_>,
-) -> Result<()> {
-	if !seen_thread_ids.insert(source.thread_id.to_owned())
-		|| run_has_terminal_thread_archive_event(state_store, source.run_id)?
-	{
-		return Ok(());
-	}
-
-	candidates.push(ThreadArchiveCandidate {
-		issue_id: source.issue_id.to_owned(),
-		issue_identifier: source.issue_identifier.to_owned(),
-		run_id: source.run_id.to_owned(),
-		attempt_number: source.attempt_number,
-		thread_id: source.thread_id.to_owned(),
-		sequence_number: source
-			.sequence_number
-			.unwrap_or(state_store.event_count(source.run_id)?.saturating_add(1)),
-	});
-
-	Ok(())
-}
-
-fn run_has_terminal_thread_archive_event(state_store: &StateStore, run_id: &str) -> Result<bool> {
-	for event_type in ["thread/archive", "thread/archive/discarded"] {
-		if state_store.run_has_protocol_event(run_id, event_type)? {
-			return Ok(true);
-		}
-	}
-
-	Ok(false)
-}
-
-fn archive_completed_issue_thread_best_effort(
-	project: &ServiceConfig,
-	state_store: &StateStore,
-	process_env: &AppServerProcessEnv,
-	transport: &str,
-	candidate: &ThreadArchiveCandidate,
-) {
-	let archive_request = AppServerThreadArchiveRequest {
-		run_id: &candidate.run_id,
-		issue_id: &candidate.issue_id,
-		attempt_number: candidate.attempt_number,
-		listen: transport,
-		process_env,
-		thread_id: &candidate.thread_id,
-		sequence_number: candidate.sequence_number,
-	};
-	#[cfg(not(test))]
-	let archive_result = agent::archive_app_server_thread_after_success(&archive_request, state_store);
-	#[cfg(test)]
-	let archive_result = {
-		state_store
-			.append_event(
-				archive_request.run_id,
-				archive_request.sequence_number,
-				"thread/archive",
-				&serde_json::json!({
-					"threadId": archive_request.thread_id,
-					"issueId": archive_request.issue_id,
-					"attemptNumber": archive_request.attempt_number,
-				})
-				.to_string(),
-			)
-			.map(|()| AppServerThreadArchiveOutcome::Archived)
-	};
-
-	match archive_result {
-		Ok(AppServerThreadArchiveOutcome::Archived) => tracing::info!(
-			project_id = project.service_id(),
-			issue_id = candidate.issue_id,
-			issue = candidate.issue_identifier,
-			run_id = candidate.run_id,
-			attempt = candidate.attempt_number,
-			thread_id = %candidate.thread_id,
-			"Archived completed issue app-server thread."
-		),
-		Ok(AppServerThreadArchiveOutcome::DiscardedMissingThread) => tracing::info!(
-			project_id = project.service_id(),
-			issue_id = candidate.issue_id,
-			issue = candidate.issue_identifier,
-			run_id = candidate.run_id,
-			attempt = candidate.attempt_number,
-			thread_id = %candidate.thread_id,
-			"Discarded completed issue app-server thread archive because the thread is missing."
-		),
-		Err(error) => tracing::warn!(
-			?error,
-			project_id = project.service_id(),
-			issue_id = candidate.issue_id,
-			issue = candidate.issue_identifier,
-			run_id = candidate.run_id,
-			attempt = candidate.attempt_number,
-			thread_id = %candidate.thread_id,
-			"Failed to archive completed issue app-server thread; leaving completed run intact."
-		),
-	}
-}
-
 fn maybe_execute_deterministic_closeout<T>(
 	tracker: &T,
 	project: &ServiceConfig,
@@ -935,93 +603,6 @@ fn write_git_credentials_operation_marker(issue_run: &IssueRunPlan) {
 		issue_run.attempt_number,
 		RUN_OPERATION_GIT_CREDENTIALS,
 	);
-}
-
-fn execute_deterministic_closeout<T>(
-	tracker: &T,
-	project: &ServiceConfig,
-	workflow: &WorkflowDocument,
-	state_store: &StateStore,
-	issue_run: &IssueRunPlan,
-	tracker_tool_bridge: &TrackerToolBridge<'_>,
-	review_context: &ReviewHandoffContext,
-) -> Result<()>
-where
-	T: IssueTracker + ?Sized,
-{
-	write_run_operation_marker_best_effort(
-		&issue_run.worktree.path,
-		&issue_run.run_id,
-		issue_run.attempt_number,
-		RUN_OPERATION_REVIEW_WRITEBACK,
-	);
-
-	let pr_url = review_context.recorded_pr_url.as_deref().ok_or_else(|| {
-		eyre::eyre!(
-			"Retained closeout run `{}` for issue `{}` requires a recorded PR URL.",
-			issue_run.run_id,
-			issue_run.issue.identifier
-		)
-	})?;
-	let pull_request = tracker_tool_bridge.validate_deterministic_closeout_pr(pr_url)?;
-	let cleanup_commit_sha = worktree_head_oid(&issue_run.worktree.path)?;
-
-	ensure_closeout_issue_completed_state(tracker, workflow, issue_run)?;
-
-	tracker_tool_bridge.apply_validated_deterministic_closeout(pull_request)?;
-
-	cleanup_completed_post_review_lane(project, workflow, state_store, issue_run)?;
-	write_cleanup_complete_lifecycle_event(
-		tracker,
-		project,
-		state_store,
-		issue_run,
-		Some(pr_url),
-		cleanup_commit_sha.as_deref(),
-	)?;
-
-	tracker_tool_bridge.clear_closeout_issue_scope()?;
-
-	Ok(())
-}
-
-fn ensure_closeout_issue_completed_state<T>(
-	tracker: &T,
-	workflow: &WorkflowDocument,
-	issue_run: &IssueRunPlan,
-) -> Result<()>
-where
-	T: IssueTracker + ?Sized,
-{
-	let tracker_policy = workflow.frontmatter().tracker();
-	let completed_state = tracker_policy.resolved_completed_state();
-	let mut refreshed_issues = tracker.refresh_issues(slice::from_ref(&issue_run.issue.id))?;
-	let current_issue = refreshed_issues.pop().unwrap_or_else(|| issue_run.issue.clone());
-
-	if current_issue.state.name == completed_state {
-		return Ok(());
-	}
-	if current_issue.state.name != tracker_policy.success_state() {
-		eyre::bail!(
-			"Retained closeout for issue `{}` requires tracker state `{}` or `{}`, but the refreshed issue is `{}`.",
-			current_issue.identifier,
-			tracker_policy.success_state(),
-			completed_state,
-			current_issue.state.name
-		);
-	}
-
-	let state_id = current_issue.state_id_for_name(completed_state).ok_or_else(|| {
-		eyre::eyre!(
-			"Issue `{}` does not expose tracker state `{}` on its team.",
-			current_issue.identifier,
-			completed_state
-		)
-	})?;
-
-	tracker.update_issue_state(&current_issue.id, state_id)?;
-
-	Ok(())
 }
 
 fn run_completion_repo_gate(
