+++ apps/decodex/src/orchestrator/execution_closeout.rs
@@ -1,0 +1,95 @@
+use std::slice;
+
+use super::{
+	IssueRunPlan, IssueTracker, RUN_OPERATION_REVIEW_WRITEBACK, Result, ReviewHandoffContext,
+	ServiceConfig, StateStore, TrackerToolBridge, WorkflowDocument,
+	cleanup_completed_post_review_lane, eyre, worktree_head_oid,
+	write_cleanup_complete_lifecycle_event, write_run_operation_marker_best_effort,
+};
+
+pub(super) fn execute_deterministic_closeout<T>(
+	tracker: &T,
+	project: &ServiceConfig,
+	workflow: &WorkflowDocument,
+	state_store: &StateStore,
+	issue_run: &IssueRunPlan,
+	tracker_tool_bridge: &TrackerToolBridge<'_>,
+	review_context: &ReviewHandoffContext,
+) -> Result<()>
+where
+	T: IssueTracker + ?Sized,
+{
+	write_run_operation_marker_best_effort(
+		&issue_run.worktree.path,
+		&issue_run.run_id,
+		issue_run.attempt_number,
+		RUN_OPERATION_REVIEW_WRITEBACK,
+	);
+
+	let pr_url = review_context.recorded_pr_url.as_deref().ok_or_else(|| {
+		eyre::eyre!(
+			"Retained closeout run `{}` for issue `{}` requires a recorded PR URL.",
+			issue_run.run_id,
+			issue_run.issue.identifier
+		)
+	})?;
+	let pull_request = tracker_tool_bridge.validate_deterministic_closeout_pr(pr_url)?;
+	let cleanup_commit_sha = worktree_head_oid(&issue_run.worktree.path)?;
+
+	ensure_closeout_issue_completed_state(tracker, workflow, issue_run)?;
+
+	tracker_tool_bridge.apply_validated_deterministic_closeout(pull_request)?;
+
+	cleanup_completed_post_review_lane(project, workflow, state_store, issue_run)?;
+	write_cleanup_complete_lifecycle_event(
+		tracker,
+		project,
+		state_store,
+		issue_run,
+		Some(pr_url),
+		cleanup_commit_sha.as_deref(),
+	)?;
+
+	tracker_tool_bridge.clear_closeout_issue_scope()?;
+
+	Ok(())
+}
+
+pub(super) fn ensure_closeout_issue_completed_state<T>(
+	tracker: &T,
+	workflow: &WorkflowDocument,
+	issue_run: &IssueRunPlan,
+) -> Result<()>
+where
+	T: IssueTracker + ?Sized,
+{
+	let tracker_policy = workflow.frontmatter().tracker();
+	let completed_state = tracker_policy.resolved_completed_state();
+	let mut refreshed_issues = tracker.refresh_issues(slice::from_ref(&issue_run.issue.id))?;
+	let current_issue = refreshed_issues.pop().unwrap_or_else(|| issue_run.issue.clone());
+
+	if current_issue.state.name == completed_state {
+		return Ok(());
+	}
+	if current_issue.state.name != tracker_policy.success_state() {
+		eyre::bail!(
+			"Retained closeout for issue `{}` requires tracker state `{}` or `{}`, but the refreshed issue is `{}`.",
+			current_issue.identifier,
+			tracker_policy.success_state(),
+			completed_state,
+			current_issue.state.name
+		);
+	}
+
+	let state_id = current_issue.state_id_for_name(completed_state).ok_or_else(|| {
+		eyre::eyre!(
+			"Issue `{}` does not expose tracker state `{}` on its team.",
+			current_issue.identifier,
+			completed_state
+		)
+	})?;
+
+	tracker.update_issue_state(&current_issue.id, state_id)?;
+
+	Ok(())
+}
